maotai_stock_financial_analysis/requirements.txt
pandas==2.1.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
matplotlib==3.8.2
baostock==0.8.8
python-dotenv>=1.0.0